rename to gpflow.utilities.traversal
