+++ gpflow/utilities/__init__.py
@@ -1,4 +1,4 @@
 from .bijectors import *
 from .misc import *
 from .multipledispatch import Dispatcher
-from .utilities import *
+from .traversal import *
