+++ gpflow/utilities/traversal.py
